feat: change return type to string
When looking into evaluating larger than Int size numbers I realized it might be useful to make the return type String instead of Int.

Also realized that I could use Char comparisons to determine if a Char digit is less than or equal to another Char digit ('0' is less than '1') instead of having to convert Char back to a digit.