feat: updated for large numbers
When testing with Int.MAX_VALUE I noticed that it was taking 6-7 seconds to complete. Also we were limited to only testing Ints.

This makes it so we can test large values just as fast as small ones but also test very large numbers. The main trick was to stop using numbers at all and just compare Chars (as '0' is < '1'). Additionally made the loop more primitive and used a StringBuilder.